% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{trace_replace}
\alias{trace_replace}
\title{trace_replace}
\usage{
trace_replace(trace, path, value)
}
\arguments{
\item{trace}{a single plotly trace}

\item{path}{character vector of path elements pointing to a trace property: c("marker", "line", "size")}

\item{value}{a value to assign to that trace property}
}
\description{
trace_replace
}
